Added code to distribute hourly credits when the other credits are being distributed in a project's retrospective. Note that the hourly items are not "part" of the retrospective per se, its just that their credits are distributed when the credits for regular issues are being distributed in a retrospective.
Also, note that (at the moment) an hourly item being accepted does not
start the retrospective nor does the hourly credits contribute to the
threshold credits for a retrospective to occur. A retrospective is
only started if the pending credits for the "regular" items reaches
the threshold or the maximum allotted time since the first unaccounted
"regular" item was created has passed.